pyproject.toml
[project]
name = "repl-nix-blackjackbattle"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "discord-py>=2.4.0",
]
